feat!: revise combinator naming and switch counts from uint to int
Closes #151

- `Prefixed` -> `Preceded` (arguments flipped to match call order)
- `Suffixed` -> `Terminated`
- `Any` -> `IsA`
- `Not` -> `IsNot`
- Fill removed (was an exact alias of Repeat)
- count parameters switched from uint to int
- `Named` added, for building a Predicate from a plain function

+++ zerowidth_test.go
@@ -128,6 +128,6 @@ func TestManyCount0ZeroWidthTerminates(t *testing.T) {
 		rem, count, err := chomp.ManyCount0(chomp.Digit0()).Run("abc")
 		require.NoError(t, err)
 		assert.Equal(t, "abc", rem)
-		assert.Equal(t, uint(0), count)
+		assert.Equal(t, 0, count)
 	})
 }
